moved directions component to its own page, reorganized contact page, fixed dir panel bug (also now shows original gmaps dir panel instead of creating custom one; yay!)
